Put Platform class into it's own file
